requirements.txt
ultralytics==8.2.0
streamlit==1.35.0
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4